Add navBar component
