+++ frontend/src/components/Map/Map.module.css
@@ -12,17 +12,6 @@
   gap: 26px;
 }
 
-.trackerContainerButtons {
-  display: flex;
-  flex-direction: row;
-  gap: 60px;
-  padding: 16px 64px;
-  background-color: white;
-  border-radius: 20px;
-  justify-content: center;
-  align-items: center;
-}
-
 .trackerButton {
   width: 80px;
   height: 50px;
