+++ frontend/src/components/NavBar/index.ts
@@ -1,0 +1,1 @@
+export * from "./NavBar";
